requirements.txt
flask
Werkzeug
tensorflow
numpy
pillow